copy files to alpha, and deprecate original
