+++ app/components/primer/alpha/navigation/tab_component.html.erb
@@ -1,0 +1,11 @@
+<%= wrapper do %>
+  <%= render Primer::BaseComponent.new(**@system_arguments) do %>
+    <%= icon %>
+    <% if text.present? %>
+      <%= text %>
+    <% else %>
+      <%= content %>
+    <% end %>
+    <%= counter %>
+  <% end %>
+<% end %>
